+++ backend/src/main/java/com/prograngers/backend/dto/solution/reqeust/ScarpSolutionPostRequest.java
@@ -1,6 +1,6 @@
 package com.prograngers.backend.dto.solution.reqeust;
 
-import com.prograngers.backend.entity.Solution;
+import com.prograngers.backend.entity.solution.Solution;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
