+++ backend/src/main/java/com/prograngers/backend/dto/solution/reqeust/SolutionPatchRequest.java
@@ -1,8 +1,8 @@
 package com.prograngers.backend.dto.solution.reqeust;
 
-import com.prograngers.backend.entity.*;
-import com.prograngers.backend.entity.constants.AlgorithmConstant;
-import com.prograngers.backend.entity.constants.DataStructureConstant;
+import com.prograngers.backend.entity.solution.AlgorithmConstant;
+import com.prograngers.backend.entity.solution.DataStructureConstant;
+import com.prograngers.backend.entity.solution.Solution;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
