+++ backend/src/main/java/com/prograngers/backend/dto/solution/response/SolutionDetailResponse.java
@@ -1,8 +1,8 @@
 package com.prograngers.backend.dto.solution.response;
 
 import com.prograngers.backend.entity.Comment;
-import com.prograngers.backend.entity.Problem;
-import com.prograngers.backend.entity.Solution;
+import com.prograngers.backend.entity.problem.Problem;
+import com.prograngers.backend.entity.solution.Solution;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
