+++ backend/src/main/java/com/prograngers/backend/dto/solution/response/SolutionListSolution.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.dto.solution.response;
 
-import com.prograngers.backend.entity.constants.AlgorithmConstant;
-import com.prograngers.backend.entity.constants.DataStructureConstant;
+import com.prograngers.backend.entity.solution.AlgorithmConstant;
+import com.prograngers.backend.entity.solution.DataStructureConstant;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
