+++ backend/src/main/java/com/prograngers/backend/dto/solution/response/SolutionUpdateFormResponse.java
@@ -1,8 +1,8 @@
 package com.prograngers.backend.dto.solution.response;
 
-import com.prograngers.backend.entity.constants.AlgorithmConstant;
-import com.prograngers.backend.entity.constants.DataStructureConstant;
-import com.prograngers.backend.entity.Solution;
+import com.prograngers.backend.entity.solution.AlgorithmConstant;
+import com.prograngers.backend.entity.solution.DataStructureConstant;
+import com.prograngers.backend.entity.solution.Solution;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
