+++ backend/src/main/java/com/prograngers/backend/entity/solution/DataStructureConstant.java
@@ -1,4 +1,4 @@
-package com.prograngers.backend.entity.constants;
+package com.prograngers.backend.entity.solution;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
