+++ backend/src/main/java/com/prograngers/backend/repository/review/ReviewRepository.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.repository.review;
 
 import com.prograngers.backend.entity.Review;
-import com.prograngers.backend.entity.Solution;
+import com.prograngers.backend.entity.solution.Solution;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
