+++ backend/src/main/java/com/prograngers/backend/repository/solution/SolutionRepository.java
@@ -1,7 +1,7 @@
 package com.prograngers.backend.repository.solution;
 
 import com.prograngers.backend.entity.member.Member;
-import com.prograngers.backend.entity.Solution;
+import com.prograngers.backend.entity.solution.Solution;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
