+++ backend/src/main/java/com/prograngers/backend/service/MemberService.java
@@ -3,8 +3,8 @@ package com.prograngers.backend.service;
 import com.prograngers.backend.dto.member.response.MemberProfileResponse;
 import com.prograngers.backend.dto.request.UpdateMemberAccountInfoRequest;
 import com.prograngers.backend.dto.response.member.MemberAccountInfoResponse;
-import com.prograngers.backend.entity.Badge;
-import com.prograngers.backend.entity.Solution;
+import com.prograngers.backend.entity.badge.Badge;
+import com.prograngers.backend.entity.solution.Solution;
 import com.prograngers.backend.entity.member.Member;
 import com.prograngers.backend.exception.badrequest.BlankNicknameException;
 import com.prograngers.backend.exception.notfound.MemberNotFoundException;
